Moves everything into the midiv1 package to represent the MIDI 1 specification
