+++ midiv1/velocity_randomizer.go
@@ -1,4 +1,4 @@
-package midi
+package midiv1
 
 import (
 	"errors"
