+++ midiv1/velocity_test.go
@@ -1,4 +1,4 @@
-package midi
+package midiv1
 
 import "testing"
